+++ NEWS.md
@@ -1,3 +1,17 @@
+# swmmr 0.7.0.9000
+
+## Breaking changes
+
+* `read_inp` sections are now all in lower letters.
+
+## New features
+
+* New functions to plot swmm objects from `*.inp` files.
+
+* `read_inp` returns now an object of class "swmm_inp".
+
+## Bug fixes
+
 # swmmr 0.7.0
 
 * Added a `NEWS.md` file to track changes to the package.